+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.4** 28/Mar/2018
+* API doc pattern compilation error on Java7 #584
 * API doc enhancement - support markdown syntax in description #579
 
 **1.8.3** 27/Mar/2018